Clean up the namedpipe_rpcss code
