+++ c/meterpreter/source/extensions/priv/namedpipe_rpcss.h
@@ -1,0 +1,6 @@
+#ifndef _METERPRETER_SOURCE_EXTENSION_PRIV_PRIV_SERVER_ELEVATE_TECHNIQUES_NAMEDPIPE_RPCSS_H
+#define _METERPRETER_SOURCE_EXTENSION_PRIV_PRIV_SERVER_ELEVATE_TECHNIQUES_NAMEDPIPE_RPCSS_H
+
+DWORD elevate_via_service_namedpipe_rpcss(Remote* remote, Packet* packet);
+
+#endif
